created on_task_updated callback for tasks
